Feature/fe wallet

+++ backend/controllers/walletController.js
@@ -27,15 +27,20 @@ const addWallet = async (req, res, next) => {
       isActiveWallet,
     } = req.body;
 
-    //   Calculate reset time
+    // Calculate reset time
     let resetTime;
-    if (resetPeriod.toLowerCase() === "monthly") {
-      resetTime = moment().add(1, "months").format();
-    } else if (resetPeriod.toLowerCase() === "weekly") {
-      resetTime = moment().add(1, "weeks").format();
+    if (resetBalance) {
+      if (resetPeriod.toLowerCase() === "monthly") {
+        resetTime = moment().add(1, "months").format();
+      } else if (resetPeriod.toLowerCase() === "weekly") {
+        resetTime = moment().add(1, "weeks").format();
+      } else {
+        resetTime = moment().add(1, "years").format();
+      }
     } else {
-      resetTime = moment().add(1, "years").format();
+      resetTime = null;
     }
+
     // create a wallet
     const wallet = await Wallet.create({
       user: req.user._id,
@@ -45,7 +50,7 @@ const addWallet = async (req, res, next) => {
       hasBudget,
       budgetAmount,
       resetBalance,
-      resetPeriod,
+      resetPeriod: resetBalance ? resetPeriod : "",
       isActiveWallet,
       activeDate: moment().format(),
       resetTime,
@@ -56,6 +61,25 @@ const addWallet = async (req, res, next) => {
 
     // update the user Model with wallet id
     if (isActiveWallet) {
+      let previousActiveWallet = await Wallet.find({
+        user: req.user._id,
+        isActiveWallet: true,
+        _id: { $nin: wallet._id },
+      });
+      if (!previousActiveWallet) {
+        return next(new HttpError("Linking the wallet with user failed", 400));
+      }
+      if (previousActiveWallet.length > 0) {
+        let updatePreviousWallet = await Wallet.findByIdAndUpdate(
+          previousActiveWallet[0]._id,
+          { isActiveWallet: false }
+        );
+        if (!updatePreviousWallet) {
+          return next(
+            new HttpError("Linking the wallet with user failed", 400)
+          );
+        }
+      }
       let userLinked = await User.findByIdAndUpdate(req.user._id, {
         activeWallet: wallet._id,
       });
@@ -69,6 +93,7 @@ const addWallet = async (req, res, next) => {
       message: "Wallet added successfully",
     });
   } catch (error) {
+    console.log(error);
     return next(new HttpError("Internal Server error", 500));
   }
 };
@@ -88,38 +113,49 @@ const editWallet = async (req, res, next) => {
       walletId,
       name,
       currency,
-      balance,
       hasBudget,
       budgetAmount,
-      resetBalance,
-      resetPeriod,
       isActiveWallet,
     } = req.body;
 
-    //   Calculate reset time
-    let resetTime;
-    if (resetPeriod.toLowerCase() === "monthly") {
-      resetTime = moment().add(1, "months").format();
-    } else if (resetPeriod.toLowerCase() === "weekly") {
-      resetTime = moment().add(1, "weeks").format();
-    } else {
-      resetTime = moment().add(1, "years").format();
+    // update the user Model with wallet id
+    if (isActiveWallet) {
+      let previousActiveWallet = await Wallet.find({
+        user: req.user._id,
+        isActiveWallet: true,
+        _id: { $nin: walletId },
+      });
+      if (!previousActiveWallet) {
+        return next(new HttpError("Linking the wallet with user failed", 400));
+      }
+      if (previousActiveWallet.length > 0) {
+        let updatePreviousWallet = await Wallet.findByIdAndUpdate(
+          previousActiveWallet[0]._id,
+          { isActiveWallet: false }
+        );
+        if (!updatePreviousWallet) {
+          return next(
+            new HttpError("Linking the wallet with user failed", 400)
+          );
+        }
+      }
+      let userLinked = await User.findByIdAndUpdate(req.user._id, {
+        activeWallet: walletId,
+      });
+      if (!userLinked) {
+        return next(new HttpError("Linking the wallet with user failed", 400));
+      }
     }
+
     // update the wallet
     const wallet = await Wallet.findByIdAndUpdate(
       walletId,
       {
-        user: req.user._id,
         name,
         currency,
-        balance,
         hasBudget,
         budgetAmount,
-        resetBalance,
-        resetPeriod,
         isActiveWallet,
-        activeDate: moment().format(),
-        resetTime,
       },
       {
         new: true,
@@ -128,22 +164,13 @@ const editWallet = async (req, res, next) => {
     if (!wallet) {
       return next(new HttpError("Wallet update failed", 400));
     }
-
-    // update the user Model with wallet id
-    if (isActiveWallet) {
-      let userLinked = await User.findByIdAndUpdate(req.user._id, {
-        activeWallet: wallet._id,
-      });
-      if (!userLinked) {
-        return next(new HttpError("Linking the wallet with user failed", 400));
-      }
-    }
     res.status(201).send({
       success: true,
       data: { wallet },
       message: "Wallet Updated successfully",
     });
   } catch (error) {
+    console.log(error);
     return next(new HttpError("Internal Server error", 500));
   }
 };
@@ -160,7 +187,18 @@ const deleteWallet = async (req, res, next) => {
       );
     }
     const { walletId } = req.params;
-
+    let walletDetails = await Wallet.findById(walletId);
+    if (!walletDetails) {
+      return next(new HttpError("Wrong wallet id", 400));
+    }
+    if (walletDetails.isActiveWallet) {
+      return next(
+        new HttpError(
+          "The wallet you are about to delete is an active wallet please select a wallet and make it active in order to delete this wallet",
+          400
+        )
+      );
+    }
     // delete the wallet
     const transaction = await Transaction.deleteMany({ wallet: walletId });
     if (!transaction) {
@@ -175,6 +213,7 @@ const deleteWallet = async (req, res, next) => {
       message: "Wallet deleted successfully",
     });
   } catch (error) {
+    console.log(error);
     return next(new HttpError("Internal Server error", 500));
   }
 };
@@ -204,7 +243,7 @@ const getWalletById = async (req, res, next) => {
 // @access  Private
 const getWalletsForUser = async (req, res, next) => {
   try {
-    let limit = 10;
+    let limit = 6;
     const { _id: userId } = req.user;
     let { page, searchString } = req.query;
     if (!page || !page.trim()) {
@@ -212,30 +251,55 @@ const getWalletsForUser = async (req, res, next) => {
     }
     if (searchString.trim()) {
       let toSkip = parseInt(page - 1) * limit;
+
+      const walletCount = await Wallet.find({
+        user: userId,
+        name: { $regex: searchString, $options: "i" },
+      }).count();
+
+      let lastPage =
+        Math.ceil(walletCount / parseInt(limit)) === 0
+          ? 1
+          : Math.ceil(walletCount / parseInt(limit));
+
       const wallets = await Wallet.find({
         user: userId,
         name: { $regex: searchString, $options: "i" },
       })
+        .sort({ isActiveWallet: -1 })
         .skip(toSkip)
         .limit(limit)
-        .select("name");
-      if (!wallets) {
+        .select(["name", "isActiveWallet", "resetPeriod"]);
+
+      if (!wallets || (!walletCount && walletCount !== 0)) {
         return next(
           new HttpError("An error occured while fetching wallets", 400)
         );
       }
+
       res.status(201).send({
         success: true,
         data: wallets,
+        lastPage,
         message: "Wallet fetched successfully",
       });
     } else {
       let toSkip = parseInt(page - 1) * limit;
       const wallets = await Wallet.find({ user: userId })
+        .sort({ isActiveWallet: -1 })
         .skip(toSkip)
         .limit(limit)
-        .select("name");
-      if (!wallets) {
+        .select(["name", "isActiveWallet", "resetPeriod"]);
+
+      const walletCount = await Wallet.find({
+        user: userId,
+      }).count();
+      let lastPage =
+        Math.ceil(walletCount / parseInt(limit)) === 0
+          ? 1
+          : Math.ceil(walletCount / parseInt(limit));
+
+      if (!wallets || (!walletCount && walletCount !== 0)) {
         return next(
           new HttpError("An error occured while fetching wallets", 400)
         );
@@ -243,6 +307,7 @@ const getWalletsForUser = async (req, res, next) => {
       res.status(201).send({
         success: true,
         data: wallets,
+        lastPage,
         message: "Wallet fetched successfully",
       });
     }
